+++ src/main/java/fi/aalto/itia/saga/prosumer/util/Scheduler.java
@@ -1,5 +1,6 @@
 package fi.aalto.itia.saga.prosumer.util;
 
+import java.math.BigDecimal;
 import java.util.Date;
 import java.util.Properties;
 
@@ -7,6 +8,12 @@ import fi.aalto.itia.saga.prosumer.util.io.MatlabIO;
 import fi.aalto.itia.saga.simulation.messages.DayAheadContentResponse;
 import fi.aalto.itia.saga.util.Utility;
 
+/**
+ * This class schedules the charging of the storage by calling the optimization
+ * 
+ * @author giovanc1
+ *
+ */
 public class Scheduler {
 
 	// TODO put this in a config file maybe
@@ -21,6 +28,7 @@ public class Scheduler {
 	private static final String FILE_IN_NAME = "FILE_IN_NAME";
 	private static final String START_ID = "START_ID";
 
+	// loading properties
 	static {
 		Properties properties = Utility.getProperties(FILE_NAME_PROPERTIES);
 		fileDir = properties.getProperty(FILE_DIR);
@@ -29,18 +37,17 @@ public class Scheduler {
 		id = Integer.parseInt(properties.getProperty(START_ID));
 	};
 
-	// TODO make it return an object which contains the OPT result
-	public static synchronized DayAheadContentResponse optimizeMatlab(int h, int r,
-			double[] k, double s0, double sh, double pmax, double[] q,
-			double w, double[] tUp, double[] tDw, double tsize,
-			Date dayAheadMidnight) {
+	public static synchronized DayAheadContentResponse optimizeMatlab(int h,
+			int r, BigDecimal[] k, BigDecimal s0, BigDecimal sh,
+			BigDecimal pmax, BigDecimal[] q, BigDecimal w, BigDecimal[] tUp,
+			BigDecimal[] tDw, BigDecimal tsize, Date dayAheadMidnight) {
 		DayAheadContentResponse opt = null;
 		String write = MatlabIO.prepareStringOut(h, r, k, s0, sh, pmax, q, w,
 				tUp, tDw, tsize, fileDir + fileInName, id);
 		MatlabIO.writeOutputFile(write, fileDir + fileOutName);
 		if (MatlabIO.watchOptResult(fileDir, fileInName))
-			opt = MatlabIO.readOptResult(fileDir + fileInName,
-					dayAheadMidnight);
+			opt = MatlabIO
+					.readOptResult(fileDir + fileInName, dayAheadMidnight);
 		id++;
 		return opt;
 	}
